+++ spec/grape/api_spec.rb
@@ -3585,10 +3585,10 @@ XML
   end
 
   context 'body' do
-    context 'false' do
+    context 'nil' do
       before do
         subject.get '/blank' do
-          body false
+          body nil
         end
       end
       it 'returns blank body' do