+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 
 #### Fixes
 
-* [#1818](https://github.com/ruby-grape/grape/pull/1818): Fixes methods defined on included modules after remountable apis - [@myxoh](https://github.com/myxoh).
+* Your contribution here.
 * [#1796](https://github.com/ruby-grape/grape/pull/1796): Fix crash when available locales are enforced but fallback locale unavailable - [@Morred](https://github.com/Morred).
 * [#1776](https://github.com/ruby-grape/grape/pull/1776): Validate response returned by the exception handler - [@darren987469](https://github.com/darren987469).
 * [#1787](https://github.com/ruby-grape/grape/pull/1787): Add documented but not implemented ability to `.insert` a middleware in the stack - [@michaellennox](https://github.com/michaellennox).